service-image-scaler/requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
filelock==3.12.4
Flask==3.0.0
fsspec==2023.9.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.1
Pillow==10.1.0
requests==2.31.0
sympy==1.12
torch==2.1.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
python-dotenv==1.0.0